requirement.txt
asyncpg
databases[postgresql]
fastapi
SQLAlchemy
uvicorn
httpx
passlib[bcrypt]==1.7.4
bcrypt<4.0
python-jose
pydantic-settings>=2.0.0
email-validator
python-dotenv
sqladmin[full]
itsdangerous>=2.0
phonenumbers